+++ src/components/CovidInfo/CovidInfoMain.js
